Staff and student info card updates for 7.2.5 p5
